adjust the usage about size_t and ssize_t *size_t equ the length of point (4 in x86 and 8 in x64) *ssize_t is same as size_t, but it is signed send() and write() use size_t as input param and ssize_t as return value

+++ source/cyCore/core/cyc_ring_buf.cpp
@@ -196,7 +196,7 @@ ssize_t RingBuf::write_socket(socket_t fd)
 		if (m_read >= m_end) m_read = 0;
 
 		//socket buf busy, try next time
-		if(len < n) {
+		if(len < (ssize_t)n) {
 			break;
 		}
 	}
